+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/GeoPointMapper.java
@@ -146,6 +146,12 @@ public class GeoPointMapper extends Mapper {
 
     /** {@inheritDoc} */
     @Override
+    public String getAnalyzer() {
+        return null;
+    }
+
+    /** {@inheritDoc} */
+    @Override
     public void addFields(Document document, Columns columns) {
 
         Double longitude = readLongitude(columns);